requirements.txt
aiohttp==3.6.2
aiohttp-json-rpc==0.12.2
aiosqlite==0.11.0
appdirs==1.4.3
apsw-wheels==3.30.1.post3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
contextlib2==0.5.5
discord.py==1.3.1
distro==1.4.0
fuzzywuzzy==0.17.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.17.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.3
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pygsheets==2.0.3
pylint==2.4.4
PyNaCl==1.3.0
python-Levenshtein-wheels==0.13.1
pytz==2019.3
PyYAML==5.3
Red-DiscordBot==3.3.0
Red-Lavalink==0.4.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
schema==0.7.1
six==1.14.0
tqdm==4.41.1
uritemplate==3.0.1
urllib3==1.25.8
uvloop==0.14.0
websockets==6.0
wrapt==1.11.2
yarl==1.4.2
